requirements.txt
gradio==4.43.0
pyyaml==6.0.2
requests==2.32.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
pypdf==5.0.1
numpy==1.26.4
tqdm==4.66.5
google-genai>=0.3.0
